Time: 72 ms (31.09%) | Memory: 31.8 MB (24.77%) - LeetSync

+++ 264-ugly-number-ii/ugly-number-ii.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int nthUglyNumber(int n) {
-        set<long> s;
-        long ans=1;
+        set<long long> s;
+        long long ans=1;
         s.insert(ans);
         for(int i=0;i<n;i++)
         {
